feat: move SQL generation tests to dedicated files
