+++ tests/unit/sql/sql.test.js
@@ -9,29 +9,6 @@ import {
     generateSQL
 } from "../../../src/utils/sql"
 
-let oneNGraph;
-let oneOneGraph;
-let nMGraph;
-let oneNGraphAndEntity;
-
-beforeEach(() => {
-    // Load fresh data before each test
-    const oneNGraphData = 
-    readFileSync(resolve(__dirname, './graphs/1-n-relation.json'), 'utf-8');
-    oneNGraph = JSON.parse(oneNGraphData);
-
-    const oneOneGraphData = 
-    readFileSync(resolve(__dirname, './graphs/1-1-relation.json'), 'utf-8');
-    oneOneGraph = JSON.parse(oneOneGraphData);
-
-    const nMGraphData = 
-    readFileSync(resolve(__dirname, './graphs/n-m-relation.json'), 'utf-8');
-    nMGraph = JSON.parse(nMGraphData);
-
-    const oneNGraphAndEntityData = 
-    readFileSync(resolve(__dirname, './graphs/1-n-relation_alone-entity.json'), 'utf-8');
-    oneNGraphAndEntity = JSON.parse(oneNGraphAndEntityData);
-});
 
 describe("Filter graph tables", () => {
     test("1:1 relation", () => {
@@ -124,111 +101,3 @@ describe("Extract table N:M relation", () => {
         expect(tables.at(2).attributes.at(1).foreign_key).toBe("Entidad 1")
     })
 })
-
-describe("Generate SQL", () => {
-    test("1:N relation", () => {
-        const sql = generateSQL(oneNGraph);
-        const expectedSQL = 
-`CREATE TABLE Entidad (
-  Atributo VARCHAR(40) PRIMARY KEY
-);
-
-CREATE TABLE Entidad_1 (
-  Atributo VARCHAR(40) PRIMARY KEY,
-  Atributo_Relacion VARCHAR(40)
-);
-
-ALTER TABLE Entidad_1 ADD CONSTRAINT FK_Atributo_Relacion FOREIGN KEY (Atributo_Relacion) REFERENCES Entidad(Atributo);
-`;
-        // Custom matcher to ignore whitespace differences
-        expect(sql.replace(/\s+/g, '')).toBe(expectedSQL.replace(/\s+/g, ''));
-    });
-
-    test("1:1 relation, 0:1-1:1", () => {
-        oneOneGraph.relations.at(0).side1.cardinality = "0:1"
-        const sql = generateSQL(oneOneGraph);
-        const expectedSQL = 
-`CREATE TABLE Entidad_1 (
-  Atributo VARCHAR(40) PRIMARY KEY
-);
-
-CREATE TABLE Entidad (
-  Atributo VARCHAR(40) PRIMARY KEY,
-  Atributo_Relacion VARCHAR(40) UNIQUE NOT NULL
-);
-
-ALTER TABLE Entidad ADD CONSTRAINT FK_Atributo_Relacion FOREIGN KEY (Atributo_Relacion) REFERENCES Entidad_1(Atributo);
-`;
-        // Custom matcher to ignore whitespace differences
-        expect(sql.replace(/\s+/g, '')).toBe(expectedSQL.replace(/\s+/g, ''));
-    });
-
-    test("N:M relation", () => {
-        const sql = generateSQL(nMGraph);
-        const expectedSQL = 
-`CREATE TABLE Entidad (
-  Atributo VARCHAR(40) PRIMARY KEY
-);
-
-CREATE TABLE Entidad_1 (
-  Atributo VARCHAR(40) PRIMARY KEY
-);
-
-CREATE TABLE Relacion (
-  Atributo_Relacion_1 VARCHAR(40),
-  Atributo_Relacion_2 VARCHAR(40),
-  Atributo VARCHAR(40),
-  PRIMARY KEY (Atributo_Relacion_1, Atributo_Relacion_2)
-);
-
-ALTER TABLE Relacion ADD CONSTRAINT FK_Atributo_Relacion_1 FOREIGN KEY (Atributo_Relacion_1) REFERENCES Entidad(Atributo);
-ALTER TABLE Relacion ADD CONSTRAINT FK_Atributo_Relacion_2 FOREIGN KEY (Atributo_Relacion_2) REFERENCES Entidad_1(Atributo);
-`;
-
-        // Custom matcher to ignore whitespace differences
-        expect(sql.replace(/\s+/g, '')).toBe(expectedSQL.replace(/\s+/g, ''));
-    });
-
-
-    test("Complete Graph", () => {
-        const sql = generateSQL(oneNGraphAndEntity)
-
-        const expectedSQL = 
-`
-CREATE TABLE Entidad_1 (
-  Atributo VARCHAR(40) PRIMARY KEY
-);
-
-CREATE TABLE Entidad_2(
-  Atributo VARCHAR(40) PRIMARY KEY,
-  Atributo_Relacion VARCHAR(40) NOT NULL
-);
-
-CREATE TABLE Entidad (
-  Atributo VARCHAR(40) PRIMARY KEY
-);
-
-ALTER TABLE Entidad_2 ADD CONSTRAINT FK_Atributo_Relacion FOREIGN KEY (Atributo_Relacion) REFERENCES Entidad_1(Atributo);
-`;
-
-        // Custom matcher to ignore whitespace differences
-        expect(sql.replace(/\s+/g, '')).toBe(expectedSQL.replace(/\s+/g, ''));
-    });
-    test("Foreign key references should use normalized table names", () => {
-        oneNGraph.entities.at(0).name = "País";
-        oneNGraph.entities.at(1).name = "Ciudad";
-
-        const sql = generateSQL(oneNGraph);
-
-        expect(sql).toContain("CREATE TABLE Pais");
-        expect(sql).toContain("REFERENCES Pais");
-    });
-
-    test("Foreign key constraints should reference the target primary key column", () => {
-        const sql = generateSQL(oneNGraph);
-
-        expect(sql).toContain(
-            "FOREIGN KEY (Atributo_Relacion) REFERENCES Entidad(Atributo)"
-        );
-    });
-});
